+++ examples/modify-voxels.rs
@@ -46,7 +46,7 @@ fn setup(mut commands: Commands, assets: Res<AssetServer>) {
         EnvironmentMapLight {
             diffuse_map: assets.load("pisa_diffuse.ktx2"),
             specular_map: assets.load("pisa_specular.ktx2"),
-            intensity: 500.0,
+            intensity: 2000.0,
             ..default()
         },
         Name::new("camera"),